==> 02_Extra_Data_Collection.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

# Data Collection

# Imports
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

import GetOldTweets3 as got
import time
import requests

## Attempt 1

# Credits (using GetOldTweets3 package): Martin Beck, 01.12.2020
# https://towardsdatascience.com/how-to-scrape-tweets-from-twitter-59287e20f0f1

# Set up variables for first Twitter pull
username = 'TxDOT'
since_date = '2016-01-01'
until_date = '2020-02-10'
max_tweets = 10000

# Create first query object
tweetCriteria = got.manager.TweetCriteria().setUsername(username).setSince(since_date).setUntil(until_date).setMaxTweets(max_tweets)

# Greate first list of all tweets
tweets = got.manager.TweetManager.getTweets(tweetCriteria)

# Create filtered list of tweet data
user_tweets = [[tweet.date, tweet.username, tweet.text, tweet.hashtags] for tweet in tweets]

# Transform list into the base tweet collection dataframe
all_tweets = pd.DataFrame(user_tweets)

# Export to dataframe
all_tweets.to_csv(f'./data/train_data/train_data_by_user/TxDOT_tweets', index=False)

# Look at first 5 tweets from 'TxDOT'
all_tweets.head()

# Look at last 5 tweets from 'TxDOT'
all_tweets.tail()

# Define list of Twitter usernames to scrape, exclusing first Twitter user ('TxDOT')
# From: https://www.txdot.gov/driver/weather/txdot-twitter-feeds.html
twitter_users = ['TxDOTAbilene', 'TxDOTAmarillo', 'TxDOTAtlanta', 'TxDOTAustin', 
                 'TxDOTBeaumont', 'TxDOTBWD', 'TxDOTBryan', 'TxDOTChildress', 'TxDOT_CRP', 
                 'TxDOTDallas', 'TxDOTELP', 'TxDOTFortWorth', 'GalvestonFerry', 'TxDOTHouston', 
                 'HoustonTranstar', 'I35Travel', 'TxDOTLaredo', 'TxDOTLubbock', 'TxDOTLufkin', 
                 'TxDOTOdessa', 'TxDOTParis', 'TxDOTPharr', 'PortA_Ferry', 'TxDOTSanAngelo', 
                 'TxDOTSanAntonio', 'TexasHighways', 'TxDOTTyler', 'TxDOTWacoPIO', 'TXDOTWF', 
                 'TxDOTYoakum', 'ImproveMopac', 'ManorExpressway', 'DFWConnector', 'DriveMidtown', 
                 'Drive360South', 'LBJexpress', 'NTExpress', 'my290Houston']

# How many Twitters are we scraping?
len(twitter_users)

# Define a function to pull remaining users' tweets from the same time period
def update_tweets(base_df, user):
    
    # Set up variables for first Twitter pull
    username = user
    since_date = '2016-01-01'
    until_date = '2020-02-10'
    max_tweets = 10000

    # Create first query object
    tweetCriteria = got.manager.TweetCriteria().setUsername(username).setSince(since_date).setUntil(until_date).setMaxTweets(max_tweets)

    # Greate first list of all tweets
    tweets = got.manager.TweetManager.getTweets(tweetCriteria)

    # Create filtered list of tweet data
    user_tweets = [[tweet.date, tweet.username, tweet.text, tweet.hashtags] for tweet in tweets]

    # Transform list into the base tweet collection dataframe
    tweets_df = pd.DataFrame(user_tweets)
    
    # Concatenate new tweets with old tweets
    updated = pd.concat([base_df, tweets_df], axis=0, ignore_index=True, sort=True)
    
    # Export the new dataframe individually
    tweets_df.to_csv(f'./data/train_data/train_data_by_user/{user}_tweets', index=False)
    
    # Print progress update
    print(f'Done scraping. We added {len(user_tweets)} new tweets. Current df shape: {updated.shape}')
    
    # Return updated dataframe of all tweets
    return updated

# Get tweets from next 14 of the total 39 twitters users and add to the all_tweets dataframe

# Set up count
i = 0

# Loop through all twitter users to scrape and add to the all_tweets dataframe, according to Twitter API limits
for user in twitter_users:

    # Count twitters as they are scraped and get the initial time
    i += 1
    t0 = time.time()
    
    # Print progress update
    print(f'Twitter #{i} now scraping...')
    
    # Get the tweets and update the dataframe
    all_tweets = update_tweets(all_tweets, user)
    
    # Wait the the remainder of 15 minutes
    print(f'Time elapsed: {time.time() - t0}')
    time.sleep((15*60) - (time.time() - t0))
    

# Export first round of data as one dataframe
all_tweets.to_csv('./data/train_data/DOT1_tweets.csv', index=False)

## Attempt 2

# Update list of Twitters still needed to be scraped
remaining_twitter_users = twitter_users[14:]
remaining_twitter_users[:3]

# Set up variables for first Twitter pull of second attempt
username = 'HoustonTranstar'
since_date = '2016-01-01'
until_date = '2020-02-10'
max_tweets = 10000

# Create first query object
tweetCriteria = got.manager.TweetCriteria().setUsername(username).setSince(since_date).setUntil(until_date).setMaxTweets(max_tweets)

# Greate first list of all tweets
tweets = got.manager.TweetManager.getTweets(tweetCriteria)

# Create filtered list of tweet data
user_tweets = [[tweet.date, tweet.username, tweet.text, tweet.hashtags] for tweet in tweets]

# Transform list into the base tweet collection dataframe
all_tweets2 = pd.DataFrame(user_tweets)

# Look at first 5 tweets from 'HoustonTranstar'
all_tweets2.head()

# Look at last 5 tweets from 'HoustonTranstar'
all_tweets2.tail()

# Continue scraping: Get tweets from next 9 of remaining 23 twitters users and add to the all_tweets dataframe

# Set up count
i = 0

# Loop through all remaining twitter users to scrape and add to the all_tweets dataframe, 
# according to Twitter API limits
for user in remaining_twitter_users[1:]:

    # Count twitters as they are scraped and get the initial time
    i += 1
    t0 = time.time()
    
    # Print progress update
    print(f'Twitter #{i} now scraping...')
    
    # Get the tweets and update the dataframe
    all_tweets2 = update_tweets(all_tweets2, user)
    
    # Wait the the remainder of 15 minutes
    print(f'Done. Time elapsed: {time.time() - t0}')
    print()
    time.sleep(15*60)
    

# Export second round of data as one dataframe
all_tweets2.to_csv('./data/train_data/DOT2_tweets.csv', index=False)

# Combine and export data

# Import first round of data
tweets1 = pd.read_csv('./data/train_data/DOT1_tweets.csv')

# Import second round of data
tweets2 = pd.read_csv('./data/train_data/DOT2_tweets.csv')

# Combine all tweets from first 25 of 38 twitters
all_tweets_final = pd.concat([tweets1, tweets2], axis=0, ignore_index=True, sort=True)
all_tweets_final.shape

# Export full dataset of all tweets from first 25 of 38 twitters
all_tweets_final.to_csv('./data/train_data/DOT_full_tweets.csv', index=False)
