==> 03_Police_Data_Collection.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Michael Daugherty | US-DSI-10 | 02.21.2020

### Collecting tweets from police departments of 30 of the most populous cities in Texas

import GetOldTweets3 as got, import pandas as pd
import time
import os

# list of police department Twitter feeds to pull from
twitter_feeds = [
    "houstonpolice",
    "SATXPolice",
    "DallasPD",
    "Austin_Police",
    "fortworthpd",
    "EPPOLICE",
    "ArlingtonPD",
    "CorpusChristiPD",
    "PlanoPoliceDept",
    "mylaredopd",
    "LubbockPolice",
    "GarlandPD",
    "irvingpd",
    "amarillopd",
    "GrandPrairiePD",
    "McKinneyPolice",
    "friscopd",
    "PasPoliceTx",
    "KilleenPD",
    "MidlandTxPD",
    "DentonPolice",
    "WacoPolice",
    "carrolltontxpd",
    "roundrockpolice",
    "abilenepd",
    "PearlandPD",
    "RichardsonTX_PD",
    "odessapolice",
    "Sugar_Land_TX",
    "beaumont_police"
]

for feed in twitter_feeds:    # A
    start_date = "2016-01-01" # B
    end_date = "2020-02-10"   # B
    max_tweets = 18_000       # C
    
    tweetCriteria = (got.
    manager.
    TweetCriteria().
    setUsername(feed).        # specify which Twitter feed to pull from (A)
    setSince(start_date).
    setUntil(end_date).       # specify date range to pull from (B)
    setMaxTweets(max_tweets)) # limit tweets to 18k to avoid getting locked out (C)
    
    tweets = got.manager.TweetManager.getTweets(tweetCriteria) # pull the tweets
    user_tweets = [[tweet.date, tweet.username, tweet.text, tweet.hashtags] for tweet in tweets] # list of the tweets + info
    tweets_df = pd.DataFrame(user_tweets) # make a DataFrame of the tweets and their associated information
    tweets_df.to_csv(f"./data/test_data/test_data_by_user/{feed}_tweets.csv") # export the DataFrame to a CSV in the project data directory
    time.sleep(15*60) # wait 15 minutes before starting the next batch to avoid getting locked out

### Creating a CSV of tweets from police departments during hurricane Harvey

files = os.listdir("./data/test_data/test_data_by_user/") # create a list of the CSV files created above

harvey_dfs = list() # create an empty list to store DataFrames of tweets from the relevant time period

for file in files:
    tweets = pd.read_csv(f"./data/project_data/{file}") # read the CSV back in as a DataFrame
    fewer_tweets = tweets[(tweets["0"] >= "2017-08-25") & (tweets["0"] <= "2017-09-03")] # select tweets from 8/25-9/2/2017
    harvey_dfs.append(fewer_tweets) # put those tweets in the list created above

harvey_tweets = pd.concat(harvey_tweets) # stack all of the DataFrames of tweets from 8/25-9/2 into a single DataFrame
harvey_tweets.to_csv("./data/test_data/PD_HH_tweets.csv") # export to a CSV in the project data directory
