==> 04_Data_Concatenation.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

# Imports
import pandas as pd

## Combine Data: Test

# Import datasets to concatenate
pd_test = pd.read_csv('./data/test_data/PD_HH_tweets.csv')
pd_test.head()

pd_test.drop(columns='Unnamed: 0', inplace=True)

dot_test = pd.read_csv('./data/test_data/TxDOT_HH_tweets.csv')
dot_test.head()

# Concatenate dataframes
test_data = pd.concat([pd_test, dot_test], axis=0, ignore_index=True, sort=True)
test_data.head()

## Export Data: Test

test_data.to_csv('./data/test_data/combined_test_data.csv')

## Combine Data: Train

# Import datasets to concatenate
pd_train = pd.read_csv('./data/train_data/PD_full_tweets.csv')
pd_train.head()

pd_train.drop(columns='Unnamed: 0', inplace=True)

dot_train = pd.read_csv('./data/train_data/DOT_full_tweets.csv')
dot_train.head()

# Look at NaNs in dot_train
dot_train.isnull().sum()

# Look at initial shape
dot_train.shape

# Create dataframe for first non-NaN rows
dot1 = dot_train.drop(columns=['one', 'two', 'three', 'zero'])
dot1.dropna(axis=0, how='all', inplace=True)
dot1.head()

# Look for nulls
dot1.isnull().sum()

# Drop all rows where column 2 is NaN
dot1 = dot1.loc[dot1['2'].notnull()]

dot1.shape

# Create dataframe for second non-NaN rows
dot2 = dot_train.drop(columns=['1', '2', '3', '0'])
dot2.dropna(axis=0, how='all', inplace=True)
dot2.head()

# Look for nulls
dot2.isnull().sum()

# Drop all rows where column 2 is NaN
dot2 = dot2.loc[dot2['two'].notnull()]

dot2.shape

# Check the shapes of the two dataframes to concatenate
print(dot1.shape)
print(dot2.shape)

# Rename dot1 columns 
dot1.rename(columns={'0': 'zero', '1': 'one', '2': 'two', '3': 'three'}, inplace=True)
dot1.head()

# Concatenate
dot_train = pd.concat([dot1, dot2], axis=0, ignore_index=True, sort=True)
dot_train.head()

# Reorder columns
dot_train = dot_train[['zero', 'one', 'two', 'three']]
dot_train.head()

pd_train.columns

# Rename columns to match pd_train
dot_train.rename(columns={'zero': '0', 'one': '1', 'two': '2', 'three': '3'}, inplace=True)
dot_train.head()

### Concatenate DataFrames

# Concatenate dataframes
train_data = pd.concat([pd_train, dot_train], axis=0, ignore_index=True, sort=True)
train_data.head()

## Export Data: Train

train_data.to_csv('./data/train_data/combined_train_data.csv')
