==> 06_Prepare_Train_Data.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

# Imports
import numpy as np
import pandas as pd

## Prepare Data

# Import train data
df = pd.read_csv('./data/train_data/clean_train_data.csv')
df.shape

df.head()

# Drop unnecessary column
df.drop(columns='Unnamed: 0', inplace=True)
df.head()

# Rename the category column
df.rename(columns={'category ': 'class'}, inplace=True)
df.head()

# Check for nulls
df.isnull().sum()

# Remove rows with nulls
df = df[df['tweet'].notnull()]

# Export Filtered Data

# Look at the original shape
df.shape

# Filer out the dates of Hurricane Harvey (tweets contained in our test data)
df_train = df[(df['dates&time'] < '2017-08-25') | (df['dates&time'] > '2017-09-02')]

# Look at the new shape (should have fewer rows)
df_train.shape

# Export data
df_train.to_csv('./data/final_train_data.csv', index=False)
