==> 07_Model_Data_1.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

## Model Data

# Imports
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier, \
                             GradientBoostingClassifier, AdaBoostClassifier
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import confusion_matrix

# from sklearn.naive_bayes import MultinomialNB
# from sklearn.neighbors import KNeighborsClassifier
# from sklearn.svm import SVC

# Revise this imports list
from nltk.corpus import stopwords
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer
from nltk.stem.porter import PorterStemmer

## Import Data

df = pd.read_csv('./data/train_data/final_train_data.csv')
df.shape

df.head()

# Check for nulls
df.isnull().sum()

# Check out the y value distribution
df['class'].value_counts(normalize=True)

Baseline Accuracy: 95%

# Define X and y
X = df['tweet']
y = df['class']

# Train tes split data
X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=.25, 
                                                    stratify=y,
                                                    random_state=42)

## Logistic Regression: CountVecotorizer

# Set up pipeline
lr_cv_pipe = Pipeline([
    ('cvec', CountVectorizer()), 
    ('lr', LogisticRegression(solver='liblinear')) 
])

# Set up pipeline parameters
lr_cv_params = {
    'cvec__stop_words': [None, 'english'],
    'cvec__ngram_range': [(1,1), (1,2), (1,3)], 
    'cvec__max_features': [100, 500], 
    'cvec__max_df': [1.0, .95], 
    'cvec__min_df': [1, .05], 
    'lr__C': [1, 1e9], 
    'lr__penalty': ['l1', 'l2'],
    'lr__max_iter': [100, 500, 1000]
}

# Set up gridsearch
lr_cv_gs = GridSearchCV(lr_cv_pipe, lr_cv_params, n_jobs=4, cv=5, verbose=1)

# Fit the gridsearch to the training data
lr_cv_gs.fit(X_train, y_train);

# What was our highest cross-val accuracy score across model hyperparameter combinations?
lr_cv_gs.best_score_

# What model hyperparameters yielded the highest accuracy score?
lr_cv_gs.best_estimator_

# What were the parameters?
lr_cv_gs.best_params_

# How does this model perform on the test set?
lr_cv_gs.score(X_test, y_test)

This model performs with high accuracy on both the train and test datasets. It is neither overfit nor underfit.

## Logistic Regression: TfidfVecotorizer

# Set up pipeline
lr_tv_pipe = Pipeline([
    ('tvec', TfidfVectorizer()), 
    ('lr', LogisticRegression(solver='liblinear')) 
])

# Set up pipeline parameters
lr_tv_params = {
    'tvec__stop_words': [None, 'english'],
    'tvec__ngram_range': [(1,1), (1,2), (1,3)], 
    'tvec__max_features': [100, 500], 
    'tvec__max_df': [1.0, .95], 
    'tvec__min_df': [1, .05], 
    'lr__C': [1, 1e9], 
    'lr__penalty': ['l1', 'l2'],
    'lr__max_iter': [100, 500, 1000]
}

# Set up gridsearch
lr_tv_gs = GridSearchCV(lr_tv_pipe, lr_tv_params, n_jobs=4, cv=5, verbose=1)

# Fit the gridsearch to the training data
lr_tv_gs.fit(X_train, y_train);

# What was our highest cross-val accuracy score across model hyperparameter combinations?
lr_tv_gs.best_score_

# What model hyperparameters yielded the highest accuracy score?
lr_tv_gs.best_estimator_

# What were the parameters?
lr_tv_gs.best_params_

# How does this model perform on the test set?
lr_tv_gs.score(X_test, y_test)

## Support Vector Machine: CountVectorizer

# Set up pipeline
svc_cv_pipe = Pipeline([
    ('cvec', CountVectorizer()), 
    ('svc', SVC(gamma='scale')) 
])

# Set up pipeline parameters
svc_cv_params = {
    'cvec__stop_words': [None, 'english'],
    'cvec__ngram_range': [(1,1), (1,2), (1,3)], 
    'cvec__max_features': [100, 250, 500],
    'svc__C': [1, 2, 3, 5.5, 7, 8.5, 10, 20, 50], 
}

# Set up gridsearch
svc_cv_gs = GridSearchCV(svc_cv_pipe, svc_cv_params, n_jobs=4, cv=5, verbose=1)

# Fit the gridsearch to the training data
svc_cv_gs.fit(X_train, y_train);

# What was our highest cross-val accuracy score across model hyperparameter combinations?
svc_cv_gs.best_score_

# What model hyperparameters yielded the highest accuracy score?
svc_cv_gs.best_estimator_

# What were the parameters?
svc_cv_gs.best_params_

# How does this model perform on the test set?
svc_cv_gs.score(X_test, y_test)

## Support Vector Machine: TfidfVectorizer

# Set up pipeline
svc_tv_pipe = Pipeline([
    ('tvec', TfidfVectorizer()), 
    ('svc', SVC(gamma='scale')) 
])

# Set up pipeline parameters
svc_tv_params = {
    'tvec__stop_words': [None, 'english'],
    'tvec__ngram_range': [(1,1), (1,2), (1,3)], 
    'tvec__max_features': [100, 250, 500],
    'svc__C': [1, 2, 3, 5.5, 7, 8.5, 10, 20, 50], 
}

# Set up gridsearch
svc_tv_gs = GridSearchCV(svc_tv_pipe, svc_tv_params, n_jobs=4, cv=5, verbose=1)

# Fit the gridsearch to the training data
svc_tv_gs.fit(X_train, y_train);

# What was our highest cross-val accuracy score across model hyperparameter combinations?
svc_tv_gs.best_score_

# What model hyperparameters yielded the highest accuracy score?
svc_tv_gs.best_estimator_

# What were the parameters?
svc_tv_gs.best_params_

# How does this model perform on the test set?
svc_tv_gs.score(X_test, y_test)

## Import Test Data

# Import test data to get project predictions
test_df = pd.read_csv('./data/test_data/combined_test_data.csv')
test_df.shape

# Look at the first 5 rows
test_df.head()

# Drop 'Unnamed: 0' and hashtage ('3') columns
test_df.drop(columns=['Unnamed: 0', '3'], inplace=True)
test_df.head()

# Rename columns
test_df.rename(columns={'0': 'dates&time', '1': 'user', '2': 'tweet'}, inplace=True)
test_df.head()

# Check for nulls
test_df.isnull().sum()

# Remove rows will null values for 'tweet' and verify changes
test_df = test_df[test_df['tweet'].notnull()]
test_df.isnull().sum()

# Export clean test data
test_df.to_csv('./data/test_data/final_test_data.csv', index=False)

## Pickle / Export Models

import pickle

### SVC / TfidfVectorizer

# Instantiate and fit svc pipeline with best hyperparameters
best_pipe = Pipeline([
    ('tvec', TfidfVectorizer(max_features=500)),
    ('svc', SVC(gamma='scale', C=7))
])
best_pipe.fit(X_train, y_train);

# Get the score on train and test data
print(f'Train score: {best_pipe.score(X_train, y_train)}')
print(f'Train cross val score: {cross_val_score(best_pipe, X_train, y_train, cv=5).mean()}')
print(f'Test score: {best_pipe.score(X_test, y_test)}')

# Get predictions on test set of train data
y_pred = best_pipe.predict(X_test)
y_pred[:5]

# Define X and y for test data
test_X = test_df['tweet']

# Get predicitons on our true test data
test_y_pred = best_pipe.predict(test_X)
test_y_pred[:5]

# Set filename and export
filename = './assets/svc_tv.sav'
pickle.dump(best_pipe, open(filename, 'wb'))

# Get predictions on our true test data using the pickled model
model = pickle.load(open('svc_tv.sav','rb'))
print(model.predict(test_X)[:5])

### Logistic Regression / CountVectorizer

# Instantiate and fit svc pipeline with best hyperparameters
best_pipe_lr = Pipeline([
    ('cvec', CountVectorizer(max_features=500, ngram_range=(1,2))),
    ('lr', LogisticRegression(solver='liblinear', penalty='l1'))
])
best_pipe_lr.fit(X_train, y_train);

# Get the score on train and test data
print(f'Train score: {best_pipe_lr.score(X_train, y_train)}')
print(f'Train cross val score: {cross_val_score(best_pipe_lr, X_train, y_train, cv=5).mean()}')
print(f'Test score: {best_pipe_lr.score(X_test, y_test)}')

# Get predictions on test set of train data
y_pred_lr = best_pipe_lr.predict(X_test)
y_pred_lr[:5]

# Define X and y for test data
test_X = test_df['tweet']

# Get predicitons on our true test data
test_y_pred_lr = best_pipe_lr.predict(test_X)
test_y_pred_lr[:5]

# Set filename and export
filename = './assets/lr_cv.sav'
pickle.dump(best_pipe_lr, open(filename, 'wb'))

# Get predictions on our true test data using the pickled model
model_lr = pickle.load(open('./assets/lr_cv.sav','rb'))
print(model_lr.predict(test_X)[:5])

## Compare Test Data Predictions

pred_comparison = list(zip(model.predict(test_X), model_lr.predict(test_X)))
pred_comparison[:5]

# How many tweets are labelled differently between the two models?
count = 0
for tup in pred_comparison:
    svc, lr = tup
    if svc != lr:
        count +=1
count

# How many tweets between the two models were labelled useful?
count = 0
for tup in pred_comparison:
    svc, lr = tup
    if svc == 1 or lr == 1:
        count +=1
count

# How many tweets were labelled useful in both models?
count = 0
for tup in pred_comparison:
    svc, lr = tup
    if svc == 1 and lr == 1:
        count +=1
count

## Model Evaluation

# Get y predictions on test set of train data
model_lr = pickle.load(open('./assets/lr_cv.sav','rb'))
y_pred = model_lr.predict(X_test)
print(model_lr.predict(X_test)[:5])

# Generate confusion matrix
cm = confusion_matrix(y_test, y_pred)
cm_df = pd.DataFrame(cm, 
             columns=['predicted not useful', 'predicted useful'], 
             index=['actual not useful', 'actual useful'])
cm_df

# Get tn, fp, fn, tp
tn, fp, fn, tp = cm.ravel()

# Calculate sensitivity
print(f'Sensitivity: {tp / (tp + fn)}')

# Calculate specificity
print(f'Specificity: {tn / (tn + fp)}')

# Calculate precision
print(f'Precision: {tp / (tp + fp)}')

# Calculate false negative rate
print(f'False negative rate: {tn / (tn + fn)}')

# Calculate accuracy
print(f'Accuracy: {(tp + tn) / (tp + tn + fp + fn)}')


## Explore Results

# Create dataframe of test set of train data
explore_df = pd.DataFrame(X_test)
explore_df = pd.concat([explore_df, y_test], axis=1)
explore_df.head()

# Add predictions column
explore_df['pred'] = y_pred
explore_df.head()

# Create a column indicating misclassified tweet
correct = []
for x, y in zip(y_test, y_pred):
    if x == y:
        correct.append(1)
    else:
        correct.append(0)
# Verify that list contains both 112 0s
count = 0
for num in correct:
    if num == 0:
        count += 1
count

# Return a dataframe of misclassified tweets
explore_df['correct'] = correct
misclassified = explore_df[explore_df['correct'] == 0]

# Set view options and look at misclassified tweets
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', -1)
misclassified
