==> 08_Model_Data_2.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Song May | US-DSI-10 | 02.21.2020

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split,cross_val_score,GridSearchCV
from sklearn.metrics import r2_score,mean_squared_error,accuracy_score
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize, RegexpTokenizer
from nltk.corpus import stopwords

df = pd.read_csv('./data/train_data/final_train_data.csv')

df.head()

df.shape

df.isna().sum()

df.dropna(inplace = True)

X = df['tweet']
y = df['class']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42, stratify=y)

lemmatizer = WordNetLemmatizer()
def word_lemmatizer(text):
    lemm_text = lemmatizer.lemmatize(text)
    return lemm_text

df["tweet"].apply(lambda x: word_lemmatizer(x))

### Basic Model

cvec = CountVectorizer(stop_words = 'english',max_df=0.90,
    min_df=3)

X_train_cvec = cvec.fit_transform(X_train)

df_cvec_train  = pd.DataFrame(X_train_cvec.toarray(),
             columns=cvec.get_feature_names())

df_cvec_train.head()

X_test_cvec = cvec.transform(X_test)

df_cvec_test  = pd.DataFrame(X_test_cvec.toarray(),
             columns=cvec.get_feature_names())

random_forest_class = RandomForestClassifier()
random_forest_class.fit(df_cvec_train, y_train)

random_forest_class.score(df_cvec_train, y_train)

random_forest_class.score(df_cvec_test, y_test)

### Modeling with Advanced Tuning.

pipe = Pipeline([('cvec',CountVectorizer()),
                     ('rf', RandomForestClassifier())
                ])
# Pipeline parameters
pipe_params = {
    'cvec__ngram_range': [(1,2)],
    'rf__n_estimators': [100,150,300,600],
    'rf__max_depth': [None,25,50,100]
}
# Instantiating a grid search
gs = GridSearchCV(pipe, 
                  param_grid=pipe_params, n_jobs = 6, cv = 5, verbose = 5) 
# Fitting my model
gs.fit(X_train, y_train)

gs.best_params_

gs.score(X_train, y_train)

gs.score(X_test, y_test)
