==> 09_Model_Data_3.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Michael Daugherty | US-DSI-10 | 02.21.2020

import numpy as np, pandas as pd
import matplotlib.pyplot as plt
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

data = pd.read_csv('./data/train_data/final_train_data.csv')
data.head()

data.isna().sum() # check for missing values

data["class"].value_counts(normalize=True)
# baseline accuracy score=94.9%

X, y = data["tweet"], data["class"] # create feature matrix and target vector

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=42)
# split data into training and test sets with similar distribution of class values to original dataset

tfidf_params = {
    "tfidf__min_df": [None, 25, 50],
    "ada__n_estimators": [None, 100, 150, 200],
    "ada__learning_rate": [0.1, 0.5, 1.0]
}
# specify parameters of TFIDF and AdaBoost parameters with which to tune the model

tfidf_pipe = Pipeline([
    ("tfidf", TfidfVectorizer()),
    ("ada", AdaBoostClassifier())
])
# combine vectorizing and modeling in a pipeline

tfidf_gs = GridSearchCV(estimator=tfidf_pipe, param_grid=tfidf_params, n_jobs=6)
# set up gridsearch

tfidf_gs.fit(X_train, y_train);
# fit the pipeline to the data using all combinations of the above parameters to obtain the best cross-val score

tfidf_gs.best_params_
# show the combination of parameters which gave the best cross-val score

tfidf_gs.best_estimator_.score(X_train, y_train)
# show the accuracy score of the model with those parameters on the training set

tfidf_gs.best_estimator_.score(X_test, y_test)
# and on the testing set

cvec_params = {
    "cvec__min_df": [None, 25, 50],
    "ada__n_estimators": [None, 100, 150, 200],
    "ada__learning_rate": [0.1, 0.5, 1.0]
}
# create a new dictionary of parameters for gridsearch using CountVectorizer rather than TFIDF

cvec_pipe = Pipeline([
    ("cvec", CountVectorizer()),
    ("ada", AdaBoostClassifier())
])
# set up a new pipeline with CountVectorizer and AdaBoost

cvec_gs = GridSearchCV(estimator=cvec_pipe, param_grid=cvec_params, n_jobs=6)
# set up gridsearch

cvec_gs.fit(X_train, y_train);

tfidf_gs.best_params_
# best parameters for CountVectorizer and AdaBoost

tfidf_gs.best_estimator_.score(X_train, y_train)
# best accuracy score on training data

tfidf_gs.best_estimator_.score(X_test, y_test)

# best accuracy score on test data
