==> 10_Model_Predictions.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

## Test Data Predictions

# Imports
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.model_selection import cross_val_score
from sklearn.metrics import confusion_matrix
import pickle

## Import Test Data

# Import test data file
df = pd.read_csv('./data/test_data/final_test_data.csv')
df.shape

df.head()

# Define X for modeling
X = df['tweet']
X.head()

## Import Model

# Import model and look at first 5 predictions
model = pickle.load(open('./assets/lr_cv.sav','rb'))
print(model.predict(X)[:5])

# Create a new column with the predictions
df['class'] = list(model.predict(X))
df.head()

# Look at useful tweets
print(df[df['class']==1].shape)
df[df['class']==1].head()

df[df['class']==1].tail()

## Export Data with Predictions

df.to_csv('./data/final_data_with_predictions.csv', index=False)
