==> 11_Get_Locations.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Michael Daugherty, Kelly Slatery, Song May | US-DSI-10 | 02.21.2020

## Get Coordinates

# Imports
import pandas as pd
import numpy as np
import re
import spacy

# Set view options
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', -1)

# Import Data

full_df = pd.read_csv('./data/final_data_with_predictions.csv')
full_df.shape

full_df.head()

# Filter Data for Useful Tweets

# Filter df for useful tweets
df = full_df[full_df['class'] == 1]
df.shape

df.head()

df.tail()

# Create Location List

# Clean tweets: Map common abbreviations to full-length word
abbrevs = {
    'SB': 'southbound',
    'NB': 'northbound', 
    'EB': 'eastbound', 
    'WB': 'westbound', 
    'St': 'Street', 
    'Rd': 'Road', 
    'E.': 'east',
    'S.': 'south',
    'W.': 'west',
    'N.': 'north',
    'Blvd': 'boulevard',
    'Pkwy': 'parkway',
    'Hwy': 'highway',
    '@': ' at '  
}

# Map 
clean_tweets = []
for tweet in df['tweet']:
    for abbrev, word in abbrevs.items():
        tweet = tweet.replace(abbrev, word)
    tweet = re.sub(r'\b(http|https):\/\/.*[^ alt]\b', '', tweet)
    clean_tweets.append(tweet)
    
clean_tweets[:4]


# Regex to get good locations from filtered locations above
locations = []

for ad in clean_tweets:
    match_1 = re.search("I\-[0-9]+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+", ad)
    match_2 = re.search("[0-9a-zA-Z]+\sand+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+", ad)
    match_3 = re.search("I[0-9]+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+", ad)
    match_4 = re.search("[a-zA-Z]+\sat+\s[0-9a-zA-Z]+\s[0-9a-zA-Z]+", ad)
    match_5 = re.search("I\-[0-9]+[0-9a-zA-Z]+\s@+[0-9a-zA-Z]+\s[0-9a-zA-Z]+", ad)
    
    if match_1: locations.append(match_1.group())
    elif match_2: locations.append(match_2.group())
    elif match_3: locations.append(match_3.group())
    elif match_4: locations.append(match_4.group())
    elif match_5: locations.append(match_5.group())
        
print(len(locations))
locations[:10]

# Add ', Texas' to each location
texas_locations = []
for loc in locations:
    loc += ', Texas'
    texas_locations.append(loc)
texas_locations

# Create and Export Locations

# Create a dataframe of the locations
locations_df = pd.DataFrame(texas_locations, columns=['location'])
locations_df.shape

locations_df.head()

# Export locations
locations_df.to_csv('./data/locations.csv', index=False)
