==> 12_Get_Coordinates.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Song May | US-DSI-10 | 02.21.2020

import pandas as pd
import spacy
import os
import numpy as np
import datetime
import warnings
import os
import geocoder
warnings.filterwarnings('ignore')
import matplotlib.pyplot as plt
%matplotlib inline

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', -1)

tweets = pd.read_csv('./data/final_data_with_predictions.csv')

tweets.head()

useful_twts = tweets[tweets['class'] == 1]

useful_twts.shape

useful_twts.head(103)

useful_twts.reset_index()

useful_twts.drop(['dates&time','user'],axis = 1,inplace = True)

useful_twts.head()

location = ['I-10 East at Airway, Texas',
 'I-10 East at Mesa, Texas',
 'I-10 East at Asarco, Texas',
 'I-10 West at Lomaland, Texas',
 'I-10 East at Sunland, Texas', 
 'I-10 West at Hawkins, Texas']

location_sample = pd.DataFrame(location)

location_sample.rename(columns={0:'locations'}, inplace =True) 

location_sample.head()

# Credit for this cell goes to team Jasmine Vasandani, Phillip Dibert, and Eric Kropf.
coordinates = []
for i in range(len(locations['locations'])):
    g = geocoder.arcgis(locations['locations'][i])
    coordinates.append(tuple(g.latlng))

location_sample['latlng'] = coordinates

location_sample.head()

import os
from bokeh.io import show,output_file
from bokeh.models import (GMapPlot, GMapOptions, ColumnDataSource,
                         Circle, DataRange1d, PanTool,
                         WheelZoomTool, BoxSelectTool)

location_sample.head()

map_options = GMapOptions(lat = 31.76263800000001, lng = -106.49052,
                         map_type = 'roadmap', zoom = 3)

#api_key = os.environ['API_KEY']

location_sample.to_csv('./data/locations_w_LL.csv')

locations = pd.read_csv('./data/locations.csv')

locations.head()

coordinates = []
for i in range(len(locations['location'])):
    g = geocoder.arcgis(locations['location'][i])
    coordinates.append(tuple(g.latlng))

locations['latlng'] = coordinates

locations.shape

locations.to_csv('./data/locations_w_LL.csv')

la_lng = pd.read_csv('./data/lat_lng.csv')

la_lng.head()
