==> 13_Get_Lat_Lng.ipynb <==
# Project 5: Optimizing Evacuation Routes using Real-Time Traffic Information

Kelly Slatery | US-DSI-10 | 02.21.2020

# Imports
import pandas as pd

# Import Data

# Import file with locations and coordinates
ll = pd.read_csv('./data/locations_w_LL.csv')
ll.shape

ll.head()

# Drop unnecessary column
ll.drop(columns='Unnamed: 0', inplace=True)
ll.head()

# Split up latitude and longitude
lats = []
lngs = []

for coords in ll['latlng']:
    coords = coords.strip().replace('(', '').replace(')', '')
    lat, lng = coords.split()
    lats.append(float(lat.strip(',')))
    lngs.append(float(lng))

# Verify that lists line up with dataframe values
print(lats[:5])
print(lngs[:5])

# Create new columns for lats and longs
ll['lat'] = lats
ll['lng'] = lngs

# Verify that columns were created
ll.head()

# Drop all non-lat/lng columns
ll.drop(columns=['location', 'latlng'], inplace=True)
print(ll.shape)
ll.head()

# Check data types (float)
ll.dtypes

## Export Latitudes / Longitudes

ll.to_csv('./data/lat_lng.csv', index=False)
